requirements.txt
django
django-tastypie


